+++ ccnode.h
@@ -51,11 +51,11 @@ struct ccnode_t : public htab_node_t<ccnode_t>, public keynode_t<uint64_t>, publ
       //
       // serialization
       //
-      u_int s_data_size(void) const;
-      u_int s_pack_data(void *buffer, u_int bufsize) const;
-      u_int s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t upcb, void *arg);
+      size_t s_data_size(void) const;
+      size_t s_pack_data(void *buffer, size_t bufsize) const;
+      size_t s_unpack_data(const void *buffer, size_t bufsize, s_unpack_cb_t upcb, void *arg);
 
-      static u_int s_data_size(const void *buffer);
+      static size_t s_data_size(const void *buffer);
 };
 
 class cc_hash_table : public hash_table<ccnode_t> {
