+++ danode.cpp
@@ -48,16 +48,16 @@ void danode_t::reset(uint64_t _nodeid)
 // serialization
 //
 
-u_int danode_t::s_data_size(void) const
+size_t danode_t::s_data_size(void) const
 {
    return datanode_t<danode_t>::s_data_size() + 
             sizeof(uint64_t) * 3 +       // hits, proctime, xfer
             s_size_of(tstamp);         // tstamp
 }
 
-u_int danode_t::s_pack_data(void *buffer, u_int bufsize) const
+size_t danode_t::s_pack_data(void *buffer, size_t bufsize) const
 {
-   u_int datasize, basesize;
+   size_t datasize, basesize;
    void *ptr = buffer;
 
    basesize = datanode_t<danode_t>::s_data_size();
@@ -77,10 +77,10 @@ u_int danode_t::s_pack_data(void *buffer, u_int bufsize) const
    return datasize;
 }
 
-u_int danode_t::s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t upcb, void *arg)
+size_t danode_t::s_unpack_data(const void *buffer, size_t bufsize, s_unpack_cb_t upcb, void *arg)
 {
    u_short version;
-   u_int datasize, basesize;
+   size_t datasize, basesize;
    const void *ptr;
 
    basesize = datanode_t<danode_t>::s_data_size();
@@ -112,7 +112,7 @@ u_int danode_t::s_unpack_data(const void *buffer, u_int bufsize, s_unpack_cb_t u
    return datasize;
 }
 
-u_int danode_t::s_data_size(const void *buffer)
+size_t danode_t::s_data_size(const void *buffer)
 {
    u_short version = s_node_ver(buffer);
